Made some i18n API public
